backend/requirements.txt
alembic==1.5.4
blinker==1.4
certifi==2020.12.5
click==7.1.2
coverage==6.5.0
ElasticMock==1.8.1
elasticsearch==7.11.0
Flask==1.1.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
flask-redoc==0.2.1
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
psycopg2==2.8.6
PyJWT==2.0.1
pytest==7.1.2
pytest-unordered==0.5.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.36.8
urllib3==1.26.3
Unidecode==1.2.0
Werkzeug==1.0.1
url-normalize==1.4.3
